requirements.txt
# Parquet Performance Testing - Python Dependencies

# Core data processing
pandas==2.2.0
pyarrow==15.0.0
duckdb==0.10.0
geopandas==0.14.3

# S3 and cloud access
boto3==1.34.34
s3fs==2024.2.0

# Spatial processing
h3==3.7.6
shapely==2.0.2
fiona==1.9.5

# Performance monitoring
psutil==5.9.8
memory-profiler==0.61.0

# Data visualization and analysis
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0
jupyter==1.0.0

# HTTP and networking
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# Development and testing
pytest==8.0.0
black==24.1.1
flake8==7.0.0

# Configuration management
python-dotenv==1.0.0
pyyaml==6.0.1